Revert 8bea8d5f151c8af0c894d3310000b181da8e98e2 and cleanup the existing code
The sampler layer is either transparent black or solid white. The transparent black parts are ignored when drawing.
Using xor with solid white is identical to inverting a pixel. GL_EQUIV is the exact opposite of xor, that is solid white causes no pixel changes and transparent black apparently too.