Flycap Integration
Added support for the Flycap2 driver. This allows for the use of FLIR/Point Grey GigE Cameras with SSL Vision.

Most parameters are controllable within the SSL Vision GUI. The automatic parameter setting of the FLIR/Point Grey cameras has been disabled as it was found to interfere with SSL Vision usage.