graphicalClient: use argv[1] as port number
Previously, graphicalClient did only use the default port number 10006, which makes it hard to
use when the portnumber is changed in ssl-vision.

After this change, it is possible to start the graphicalClient using the port number 10010 by
./graphicalClient 10010.
This makes it easer to use in an environment, where multiple ssl-vision are running
in the same local network.

Signed-off-by: Tobias Heineken <tobias.heineken@robotics-erlangen.de>